ARMv7 bf16 basic-non-support

+++ include/tensor/operators/gather/gather.h
@@ -4,7 +4,7 @@
 inline Tensor Tensor::operator[](const size_t index) {
     std::vector<size_t> new_shape;
     size_t skipdata = get_dtype_bytes(dtype);
-    for (int i = 1; i < shape.size(); i++) {
+    for (size_t i = 1; i < shape.size(); i++) {
         new_shape.push_back(shape[i]);
         skipdata *= shape[i];
     }
@@ -21,8 +21,8 @@ inline Tensor Tensor::gather(std::vector<std::vector<size_t>> indices, Tensor ou
         out = Tensor({indices.size(), indices[0].size(), shape[1]}, dtype, out.device, device_id);
     }
 
-    for (int i = 0; i < indices.size(); i++) {
-        for (int j = 0; j < indices[0].size(); j++) {
+    for (size_t i = 0; i < indices.size(); i++) {
+        for (size_t j = 0; j < indices[0].size(); j++) {
             if (indices[i][j] >= shape[0]) {
                 throw std::invalid_argument("Index out of range");
             }
